fix hall and cinema controllers, edit datatype in seattype id

+++ Cinema.SQL/dbo/Tables/SeatTypes.sql
@@ -1,6 +1,8 @@
 ﻿CREATE TABLE [dbo].[SeatTypes] (
-    [Id]   TINYINT       IDENTITY (1, 1) NOT NULL,
+    [Id]   SMALLINT      IDENTITY (1, 1) NOT NULL,
     [Name] NVARCHAR (25) NOT NULL,
     CONSTRAINT [PK_SeatTypes] PRIMARY KEY CLUSTERED ([Id] ASC)
 );
+
+
